Add AbstractInOutDataManager to handle properly the two parts of the Database (AdvancedDataRetriever and AbstractResultsManager)
