+++ database/src/main/java/ch/heiafr/isc/datacockpit/database/AbstractInOutDataManager.java
@@ -1,0 +1,32 @@
+package ch.heiafr.isc.datacockpit.database;/*
+        This file is part of one of the Data-Cockpit libraries.
+
+        Copyright (C) 2024 EPFL
+
+        The Data-Cockpit program is free software, you can redistribute it and/or modify
+        it under the terms of GNU Lesser General Public License as published by
+        the Free Software Foundation, either version 3 of the License, or
+        (at your option) any later version.
+
+        The Data-Cockpit program is distributed in the hope that it will be useful,
+        but WITHOUT ANY WARRANTY; without even the implied warranty of
+        MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+        GNU General Public License for more details.
+
+        You should have received a copy of the GNU Lesser General Public License
+        along with this program (root directory).  If not, see <https://www.gnu.org/licenses/>.
+          
+        Author - Sébastien Rumley (sebastien.rumley@hefr.ch)
+        
+        This publication is made with the authorization of the EPFL, 
+        the institution where the author was originally employed.
+        The author is currently affiliated with the HEIA-FR.
+      
+        Contributors :
+    
+*/
+
+import ch.heiafr.isc.datacockpit.general_libraries.results.AbstractResultsManager;
+import ch.heiafr.isc.datacockpit.general_libraries.results.AdvancedDataRetriever;
+
+public abstract class AbstractInOutDataManager implements AdvancedDataRetriever, AbstractResultsManager { }
