+++ general_libraries/src/main/java/ch/heiafr/isc/datacockpit/general_libraries/results/AbstractResultsManager.java
@@ -24,33 +24,14 @@
  * 
  * Contributor list -
  */
-package ch.heiafr.isc.datacockpit.database;
+package ch.heiafr.isc.datacockpit.general_libraries.results;
 
-import java.io.Serializable;
+public interface AbstractResultsManager {
 
+	public void addExecution(Execution e);
 
+    public void clear();
 
+	public void addDataPoint(DataPoint dp);
 
-public class ResultProperty extends Property implements Serializable {
-	private static final long serialVersionUID = 1L;
-
-	public ResultProperty(String name, String value) {
-		super(name, value);
-	}
-
-	public ResultProperty(String name, double value) {
-		super(name, value+"");
-	}
-
-	public ResultProperty(String name, String value, String unit) {
-		super(name, value, unit);
-	}
-
-	public ResultProperty(String name, double value, String unit) {
-		super(name,value+"", unit);
-	}
-
-	public ResultProperty(String name, boolean b) {
-		super(name, b+"");	
-	}
 }
