+++ general_libraries/src/main/java/ch/heiafr/isc/datacockpit/general_libraries/results/ResultProperty.java
@@ -24,14 +24,33 @@
  * 
  * Contributor list -
  */
-package ch.heiafr.isc.datacockpit.database;
+package ch.heiafr.isc.datacockpit.general_libraries.results;
 
-public interface AbstractResultsManager {
+import java.io.Serializable;
 
-	public void addExecution(Execution e);
 
-    public void clear();
 
-	public void addDataPoint(DataPoint dp);
 
+public class ResultProperty extends Property implements Serializable {
+	private static final long serialVersionUID = 1L;
+
+	public ResultProperty(String name, String value) {
+		super(name, value);
+	}
+
+	public ResultProperty(String name, double value) {
+		super(name, value+"");
+	}
+
+	public ResultProperty(String name, String value, String unit) {
+		super(name, value, unit);
+	}
+
+	public ResultProperty(String name, double value, String unit) {
+		super(name,value+"", unit);
+	}
+
+	public ResultProperty(String name, boolean b) {
+		super(name, b+"");	
+	}
 }
