+++ tree/src/main/java/ch/heiafr/isc/datacockpit/tree/experiment_aut/Experiment.java
@@ -28,7 +28,7 @@ package ch.heiafr.isc.datacockpit.tree.experiment_aut;
 
 import ch.heiafr.isc.datacockpit.tree.clazzes.ClassRepository;
 import ch.heiafr.isc.datacockpit.general_libraries.results.AbstractResultsDisplayer;
-import ch.heiafr.isc.datacockpit.database.AbstractResultsManager;
+import ch.heiafr.isc.datacockpit.general_libraries.results.AbstractResultsManager;
 
 public interface Experiment {
